BC-12022 - Refactor import paths for modules to improve clarity and maintainability
