+++ src/apps/files-storage-consumer.app.ts
@@ -3,7 +3,7 @@
 import { NestFactory } from '@nestjs/core';
 
 // register source-map-support for debugging
-import { FilesStorageAMQPModule } from '@modules/files-storage-app';
+import { FilesStorageAMQPModule } from '@modules/files-storage-app/files-storage-amqp.app.module';
 import { install as sourceMapInstall } from 'source-map-support';
 
 async function bootstrap(): Promise<void> {
