+++ src/apps/preview-generator-consumer.app.ts
@@ -1,6 +1,6 @@
 /* istanbul ignore file */
 /* eslint-disable no-console */
-import { PreviewGeneratorAMQPModule } from '@modules/files-preview-app';
+import { PreviewGeneratorAMQPModule } from '@modules/files-preview-app/files-preview-amqp.app.module';
 import { NestFactory } from '@nestjs/core';
 import { install as sourceMapInstall } from 'source-map-support';
