+++ src/infra/metrics/index.ts
@@ -5,8 +5,10 @@
  * Do not export *.app.module.ts here; import them directly. *
  *********************************************************** */
 
-export { CurrentDownloadMetricsInterceptor } from './interceptor/current-download-metrics.interceptor';
-export { CurrentUploadMetricsInterceptor } from './interceptor/current-upload-metrics.interceptor';
-export { ResponseTimeMetricsInterceptor } from './interceptor/response-time-metrics.interceptor';
+export {
+	CurrentDownloadMetricsInterceptor,
+	CurrentUploadMetricsInterceptor,
+	ResponseTimeMetricsInterceptor,
+} from './interceptor';
 export { MetricsModule } from './metrics.module';
 export { MetricsService } from './metrics.service';
