+++ src/infra/preview-generator/index.ts
@@ -5,8 +5,7 @@
  * Do not export *.app.module.ts here; import them directly. *
  *********************************************************** */
 
-export { PreviewFileOptions, PreviewOptions, PreviewResponseMessage } from './interface/preview';
-export { PreviewInputMimeTypes } from './interface/preview-input-mime-types.enum';
+export { PreviewFileOptions, PreviewInputMimeTypes, PreviewOptions, PreviewResponseMessage } from './interface';
 export { PreviewGeneratorConsumerModule } from './preview-generator-consumer.module';
 export { PreviewGeneratorProducerModule } from './preview-generator-producer.module';
 export { PreviewProducer } from './preview.producer';
